Implement early firebase setup code
In this commit I found the features of firebase that were relevant to the project and added them to the firebase.js app. I also npm install'ed firebase, and some other tools which is why the package stuff got changed. I also created the firebase project on the CLI on the firebase website which will be our main interface for the backend of the project.