Update addStudent function for user creation
In this commit I used the admin.auth() library to call upon admin.auth.createUser() to create a user in the database. This is different from the previous version as the last version of the function did not create a user in firebase and make use of its authentication tools, it simply added the user's credentials to the firestore database.